+++ src/platform_impl/windows/window.rs
@@ -477,7 +477,12 @@ impl Window {
     let window = self.window.clone();
     let window_state = Arc::clone(&self.window_state);
 
+    let is_minimized = self.is_minimized();
+
     self.thread_executor.execute_in_thread(move || {
+      WindowState::set_window_flags_in_place(&mut window_state.lock(), |f| {
+        f.set(WindowFlags::MINIMIZED, is_minimized)
+      });
       WindowState::set_window_flags(window_state.lock(), window.0, |f| {
         f.set(WindowFlags::MINIMIZED, minimized)
       });
@@ -504,8 +509,7 @@ impl Window {
 
   #[inline]
   pub fn is_minimized(&self) -> bool {
-    let window_state = self.window_state.lock();
-    window_state.window_flags.contains(WindowFlags::MINIMIZED)
+    unsafe { IsIconic(self.hwnd()) }.as_bool()
   }
 
   #[inline]